parse req instead of context, refactor
